chore: rename 'index.ts' -> 'main.ts'
